Make sure performance is up to our latest stabilization point in main
(release/3.4 branch start)

This is a redoing of PR https://github.com/erigontech/erigon/pull/19663
which I accidentally squashed merged and messed the history.

**MUST BE DONE AS MERGE COMMIT**

+++ cmd/utils/app/import_cmd.go
@@ -255,7 +255,7 @@ func InsertChain(ethereum *eth.Ethereum, chain *blockgen.ChainPack, setHead bool
 		insertedBlocks[block.NumberU64()] = struct{}{}
 	}
 
-	chainRW := chainreader.NewChainReaderEth1(ethereum.ChainConfig(), direct.NewExecutionClientDirect(ethereum.ExecutionModule()), uint64(time.Hour))
+	chainRW := chainreader.NewChainReaderEth1(ethereum.ChainConfig(), direct.NewExecutionClientDirect(ethereum.ExecutionModule()), time.Hour)
 
 	ctx := context.Background()
 	if err := chainRW.InsertBlocksAndWait(ctx, chain.Blocks); err != nil {
